Initial commit.
This change contributes the `chat.py` Python script for interacting
with OpenAI models via the chat API. This is useful for users seeking
to use GPT models with a streamlined workflow for bringing their data
to the model and saving their conversations in a manner that's
accessible but without having to store their data on the platform.
According to the OpenAI documentation, customer data is not stored for
training models when using the OpenAI API.

Some examples of how to use the tool have been contributed in the
README.md file. While there are several sub-commands available, the
most interesting usage is via the `/load` and the `/save` commands.